Refactored to get the User ID from JWT instead of Request Body

+++ Core/Application/Features/Events/DeleteEvent/DeleteEventCommandHandler.cs
@@ -3,8 +3,10 @@
 // </copyright>
 namespace EventManagementAPI.Core.Application.Features.Events.DeleteEvent
 {
+    using EventManagementAPI.Core.Application.Contracts.Identity;
     using EventManagementAPI.Core.Application.Contracts.Messaging.Commands;
     using EventManagementAPI.Core.Application.Contracts.Persistence;
+    using EventManagementAPI.Core.Application.Extensions;
     using EventManagementAPI.Core.Application.Response;
     using EventManagementAPI.Core.Domain.Entities;
     using EventManagementAPI.Core.Domain.Errors;
@@ -12,22 +14,59 @@ namespace EventManagementAPI.Core.Application.Features.Events.DeleteEvent
     public class DeleteEventCommandHandler : ICommandHandler<DeleteEventCommand, Result<string>>
     {
         private readonly IRepository<Event> eventRepository;
+        private IUserService userService;
+        private readonly IUnitOfWork unitOfWork;
 
-        public DeleteEventCommandHandler(IRepository<Event> eventRepository)
+        public DeleteEventCommandHandler(IRepository<Event> eventRepository, IUserService userService, IUnitOfWork unitOfWork)
         {
             this.eventRepository = eventRepository;
+            this.userService = userService;
+            this.unitOfWork = unitOfWork;
         }
 
         public async Task<Result<string>> Handle(DeleteEventCommand request, CancellationToken cancellationToken)
         {
-            var eventDetails = await this.eventRepository.GetByIdAsync(request.Id);
-            if (eventDetails == null)
+            try
             {
-                return Result<string>.Failure(DomainErrors.Event.NotFound(request.Id));
-            }
+                var userIdString = request.User.GetUserId();
+                if (string.IsNullOrEmpty(userIdString) || !Guid.TryParse(userIdString, out var userId))
+                {
+                    return Result<string>.Failure(DomainErrors.Auth.NotAuthenticated());
+                }
 
-            await this.eventRepository.DeleteAsync(request.Id);
-            return Result<string>.Success("Removed the event successfully");
+                request.UserId = userId;
+
+                await this.unitOfWork.BeginTransactionAsync();
+
+                var user = await this.userService.GetUserDetailsByIdAsync(request.UserId);
+                if (user == null)
+                {
+                    return Result<string>.Failure(DomainErrors.Auth.UserNotFound(request.UserId));
+                }
+
+                var eventDetails = await this.eventRepository.GetByIdAsync(request.Id);
+                if (eventDetails == null)
+                {
+                    return Result<string>.Failure(DomainErrors.Event.NotFound(request.Id));
+                }
+
+                if (eventDetails.CreatedBy != request.UserId)
+                {
+                    return Result<string>.Failure(DomainErrors.Event.NotCreatedByUser(request.UserId));
+                }
+
+                await this.eventRepository.DeleteAsync(request.Id);
+
+                await this.unitOfWork.SaveChangesAsync(cancellationToken);
+                await this.unitOfWork.CommitAsync();
+
+                return Result<string>.Success("Removed the event successfully");
+            }
+            catch (Exception)
+            {
+                await this.unitOfWork.RollbackAsync();
+                return Result<string>.Failure(DomainErrors.Transaction.TransactionFailed());
+            }
         }
     }
 }
